+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This file documents all updates and releases to the ResearchNow Data Ingest pipe
 - Support for ISO3 language codes in the language filter
 ### Fixed
 - GET method for the search endpoint when using sort and filter options
+- Handle database errors in a more logical and clear way
 
 ## [0.0.5] - 2020-06-05
 ### Added